*** empty log message ***
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@495 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/run-time/bits.c
@@ -489,7 +489,7 @@ long s;
 
 public char *b_rotate(bit, s)
 char *bit;
-int s;
+long s;
 {
 	/* Rotates `bit' by `s' positions. If s is positive, rotation is done to
 	 * the right, otherwise to the left. A null value does nothing.
@@ -506,7 +506,7 @@ int s;
 
 private char *b_right_rotate(bit, s)
 char *bit;
-int s;
+long s;
 {
 	/* Rotates `bit' by `s' positions to the right */
